header: move pb into extheader
