+++ service/header/p2p_exchange_test.go
@@ -14,7 +14,7 @@ import (
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 
 	extheader "github.com/celestiaorg/celestia-node/service/header/extHeader"
-	header_pb "github.com/celestiaorg/celestia-node/service/header/pb"
+	header_pb "github.com/celestiaorg/celestia-node/service/header/extheader/pb"
 	"github.com/celestiaorg/go-libp2p-messenger/serde"
 )
