+++ service/header/p2p_server.go
@@ -9,7 +9,7 @@ import (
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 
 	extheader "github.com/celestiaorg/celestia-node/service/header/extHeader"
-	pb "github.com/celestiaorg/celestia-node/service/header/pb"
+	pb "github.com/celestiaorg/celestia-node/service/header/extheader/pb"
 	"github.com/celestiaorg/go-libp2p-messenger/serde"
 )
